\name{Crsk}
\alias{Crsk}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Create a competing risks object.
}
\description{
Combines observed time and cause indicator to be a competing risks object. For internal use only.
}
\usage{
Crsk(t, ic)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{t}{A vector of observed times.}
  \item{ic}{A vector of cause indicators.}
}
